pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llama_api_server"
version = "0.3.4"
description = "A OpenAI API compatible REST server for llama."
authors = [
    {name = "iaalm", email= "iaalmsimon@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"

classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "llamacpp>=0.1.11",
  "Flask>=2.0.0",
  "numpy",
  "pyyaml",
]

[project.optional-dependencies]
pyllama = [
    "pyllama>=0.0.9",
    "transformers>=4.27.4",
    "gptq",
]

[project.urls]
homepage = "https://github.com/iaalm/llama-api-server"
repository = "https://github.com/iaalm/llama-api-server"
